Extract details form into its own component
