+++ components/Questionnaire/enhancers.js
@@ -142,7 +142,7 @@ export const withAnswerMutation = graphql(submitAnswerMutation, {
         update: (proxy, { data: { submitAnswer } }) => {
           const queryObj = {
             getQuestionnaire,
-            variables: { slug: router.query.slug }
+            variables: { slug: router && router.query.slug }
           }
           const data = proxy.readQuery(queryObj)
           const questionIx = data.questionnaire.questions.findIndex(
